[0162] Bugfix & new OSPF command Fix a bug when receive \n char Implement stub command to declare router totally stub

+++ daemon_src/route.h
@@ -124,7 +124,7 @@ unsigned short ospf_redistrib_conn_metric;
 unsigned short ospf_redistrib_default;
 unsigned short ospf_redistrib_default_type;
 unsigned short ospf_redistrib_default_metric;
-unsigned short ospf_stub_router; // @ TODO
+uint8_t ospf_stub_router;
 unsigned int ospf_delay_timer;
 unsigned int ospf_holdtime_timer;
 char* ospf_router_id;
